+++ py/test/test_from_ngff_zarr_tensorstore.py
@@ -4,7 +4,6 @@ import tempfile
 
 import pytest
 from dask_image import imread
-
 from ngff_zarr import (
     from_ngff_zarr,
     to_multiscales,