+++ py/test/test_ngff_image_to_itk_image.py
@@ -3,14 +3,14 @@
 import itk
 import itkwasm
 import numpy as np
-from ngff_zarr import itk_image_to_ngff_image, ngff_image_to_itk_image, from_ngff_zarr
+from ngff_zarr import from_ngff_zarr, itk_image_to_ngff_image, ngff_image_to_itk_image
 
 from ._data import test_data_dir
 
 rng = np.random.default_rng(12345)
 
 
-def test_2d_itk_image(input_images):  # noqa: ARG001
+def test_2d_itk_image(input_images):
     itk_image = itk.imread(test_data_dir / "input" / "cthead1.png")
     ngff_image = itk_image_to_ngff_image(itk_image)
     itk_image_back = ngff_image_to_itk_image(ngff_image, wasm=False)
@@ -18,7 +18,7 @@ def test_2d_itk_image(input_images):  # noqa: ARG001
     assert np.sum(np.asarray(diff)) == 0.0
 
 
-def test_2d_rgb_itk_image(input_images):  # noqa: ARG001
+def test_2d_rgb_itk_image(input_images):
     array = rng.integers(0, 255, size=(224, 224, 3), dtype=np.uint8)
     itk_image = itk.image_from_array(array, is_vector=True)
     ngff_image = itk_image_to_ngff_image(itk_image)  # noqa: F841
@@ -28,7 +28,7 @@ def test_2d_rgb_itk_image(input_images):  # noqa: ARG001
     # assert np.sum(np.asarray(diff)) == 0.0
 
 
-def test_2d_itk_vector_image(input_images):  # noqa: ARG001
+def test_2d_itk_vector_image(input_images):
     array = rng.random(size=(224, 224, 3), dtype=np.float32)
     itk_image = itk.image_from_array(array, is_vector=True)
     ngff_image = itk_image_to_ngff_image(itk_image)  # noqa: F841
@@ -37,7 +37,7 @@ def test_2d_itk_vector_image(input_images):  # noqa: ARG001
     # assert np.array_equal(itk.array_from_image(itk_image), itk.array_from_image(itk_image_back))
 
 
-def test_3d_itk_image(input_images):  # noqa: ARG001
+def test_3d_itk_image(input_images):
     array = rng.integers(0, 255, size=(32, 32, 32), dtype=np.uint8)
     itk_image = itk.image_from_array(array, is_vector=False)
     ngff_image = itk_image_to_ngff_image(itk_image)
@@ -46,7 +46,7 @@ def test_3d_itk_image(input_images):  # noqa: ARG001
     assert np.sum(np.asarray(diff)) == 0.0
 
 
-def test_3d_itk_vector_image(input_images):  # noqa: ARG001
+def test_3d_itk_vector_image(input_images):
     array = rng.random(size=(224, 224, 128, 3), dtype=np.float32)
     itk_image = itk.image_from_array(array, is_vector=True)
     ngff_image = itk_image_to_ngff_image(itk_image)  # noqa: F841
@@ -55,7 +55,7 @@ def test_3d_itk_vector_image(input_images):  # noqa: ARG001
     # assert np.array_equal(itk.array_from_image(itk_image), itk.array_from_image(itk_image_back))
 
 
-def test_2d_itkwasm_image(input_images):  # noqa: ARG001
+def test_2d_itkwasm_image(input_images):
     itk_image = itk.imread(test_data_dir / "input" / "cthead1.png")
     itk_image_dict = itk.dict_from_image(itk_image)
     itkwasm_image = itkwasm.Image(**itk_image_dict)
@@ -66,7 +66,7 @@ def test_2d_itkwasm_image(input_images):  # noqa: ARG001
     )
 
 
-def test_t_index(input_images):  # noqa: ARG001
+def test_t_index(input_images):
     dataset_name = "13457537"
     store_path = test_data_dir / "input" / f"{dataset_name}.zarr"
     multiscales = from_ngff_zarr(store_path)
@@ -93,7 +93,7 @@ def test_t_index(input_images):  # noqa: ARG001
     assert itk_image.data.shape == (12, 223, 198, 6)
 
 
-def test_c_index(input_images):  # noqa: ARG001
+def test_c_index(input_images):
     dataset_name = "13457537"
     store_path = test_data_dir / "input" / f"{dataset_name}.zarr"
     multiscales = from_ngff_zarr(store_path)