+++ py/test/test_write_store_to_zip.py
@@ -9,9 +9,8 @@ import numpy as np
 import packaging.version
 import pytest
 import zarr
-
 from ngff_zarr import to_multiscales, to_ngff_image, to_ngff_zarr, write_store_to_zip
-from ngff_zarr.hcs import to_hcs_zarr, HCSPlate, from_hcs_zarr
+from ngff_zarr.hcs import HCSPlate, from_hcs_zarr, to_hcs_zarr
 from ngff_zarr.rfc9_zip import read_ozx_version
 from ngff_zarr.v04.zarr_metadata import (
     Plate,